Reorganize the options and components directories (#900)
* Rename the subscription row component to SubscriptionTableRow


* Move cross-section options files into options/shared


* Move style-only CSS files from components to styles


---------
