+++ src/scripts/block-dialog/index.tsx
@@ -1,4 +1,4 @@
-import "../components/theme.css";
+import "../styles/theme.css";
 import "./baseline.css";
 import { Button } from "@base-ui/react/button";
 import cog from "@mdi/svg/svg/cog.svg";
@@ -8,17 +8,12 @@ import * as punycode from "punycode/";
 import React, { useId, useMemo, useRef, useState } from "react";
 import { createRoot, type Root } from "react-dom/client";
 import icon from "../../icons/icon.svg";
-import buttonStyles from "../components/button.module.css";
 import dialogStyles from "../components/dialog.module.css";
 import { Dialog } from "../components/dialog.tsx";
 import { EmbeddedDialog } from "../components/embedded-dialog.tsx";
-import iconButtonStyles from "../components/icon-button.module.css";
-import labelStyles from "../components/label.module.css";
 import { MenuItem } from "../components/menu.tsx";
-import rowStyles from "../components/row.module.css";
 import { SplitButton } from "../components/split-button.tsx";
 import { SvgIcon } from "../components/svg-icon.tsx";
-import textareaStyles from "../components/textarea.module.css";
 import { browser } from "../shared/browser.ts";
 import type {
   InteractiveRuleset,
@@ -28,6 +23,11 @@ import type {
 import { translate } from "../shared/locales.ts";
 import { getRegistrableDomain } from "../shared/registrable-domain.ts";
 import type { DialogTheme, MessageName0 } from "../shared/types.ts";
+import buttonStyles from "../styles/button.module.css";
+import iconButtonStyles from "../styles/icon-button.module.css";
+import labelStyles from "../styles/label.module.css";
+import rowStyles from "../styles/row.module.css";
+import textareaStyles from "../styles/textarea.module.css";
 import styles from "./index.module.css";
 
 // ---------------------------------------------------------------------------
