+++ src/scripts/options/index.tsx
@@ -1,6 +1,6 @@
-import containerStyles from "../components/container.module.css";
-import "../components/theme.css";
-import "../components/baseline.css";
+import containerStyles from "../styles/container.module.css";
+import "../styles/theme.css";
+import "../styles/baseline.css";
 import { createRoot } from "react-dom/client";
 import { AutoThemeProvider } from "../components/theme.tsx";
 import { browser } from "../shared/browser.ts";
@@ -10,7 +10,7 @@ import { AboutSection } from "./about-section.tsx";
 import { AppearanceSection } from "./appearance-section.tsx";
 import { BackupRestoreSection } from "./backup-restore-section.tsx";
 import { GeneralSection } from "./general-section.tsx";
-import { initializePlatform } from "./platform.ts";
+import { initializePlatform } from "./shared/platform.ts";
 import {
   type OptionsQuery,
   SubscriptionSection,
