+++ src/scripts/options/shared/ruleset-editor.tsx
@@ -1,4 +1,4 @@
-import { Editor, type EditorProps } from "../components/editor.tsx";
+import { Editor, type EditorProps } from "../../components/editor.tsx";
 import { ruleset } from "./ruleset-lang.ts";
 
 export type RulesetEditorProps = Omit<EditorProps, "language">;
