+++ src/scripts/options/shared/set-interval-item.module.css
