+++ src/scripts/options/shared/set-interval-item.tsx
@@ -1,12 +1,12 @@
 import clsx from "clsx";
 import dayjs from "dayjs";
-import labelStyles from "../components/label.module.css";
-import rowStyles from "../components/row.module.css";
-import "../shared/dayjs-locales.ts";
-import { Select, SelectOption } from "../components/select.tsx";
-import { saveToLocalStorage } from "../shared/local-storage.ts";
-import { translate } from "../shared/locales.ts";
-import { storageStore } from "../shared/storage-store.ts";
+import labelStyles from "../../styles/label.module.css";
+import rowStyles from "../../styles/row.module.css";
+import "../../shared/dayjs-locales.ts";
+import { Select, SelectOption } from "../../components/select.tsx";
+import { saveToLocalStorage } from "../../shared/local-storage.ts";
+import { translate } from "../../shared/locales.ts";
+import { storageStore } from "../../shared/storage-store.ts";
 import localStyles from "./set-interval-item.module.css";
 
 export type IntervalItemKey = "syncInterval" | "updateInterval";
