+++ src/scripts/options/subscription-section/rename-dialog.tsx
@@ -2,7 +2,6 @@ import { Button } from "@base-ui/react/button";
 import { Input } from "@base-ui/react/input";
 import clsx from "clsx";
 import { useId, useState } from "react";
-import buttonStyles from "../../components/button.module.css";
 import {
   Dialog,
   DialogBody,
@@ -10,12 +9,13 @@ import {
   DialogHeader,
   DialogTitle,
 } from "../../components/dialog.tsx";
-import inputStyles from "../../components/input.module.css";
-import labelStyles from "../../components/label.module.css";
-import rowStyles from "../../components/row.module.css";
 import { translate } from "../../shared/locales.ts";
 import { sendMessage } from "../../shared/messages.ts";
 import type { Subscription, SubscriptionId } from "../../shared/types.ts";
+import buttonStyles from "../../styles/button.module.css";
+import inputStyles from "../../styles/input.module.css";
+import labelStyles from "../../styles/label.module.css";
+import rowStyles from "../../styles/row.module.css";
 
 function RenameForm({
   close,
