+++ src/scripts/options/subscription-section/table-row.tsx
@@ -2,11 +2,7 @@ import { Checkbox } from "@base-ui/react/checkbox";
 import clsx from "clsx";
 import dayjs from "dayjs";
 import { useState } from "react";
-import badgeStyles from "../../components/badge.module.css";
-import styles from "../../components/checkbox.module.css";
-import labelStyles from "../../components/label.module.css";
 import { Menu, MenuItem } from "../../components/menu.tsx";
-import tableStyles from "../../components/table.module.css";
 import { translate } from "../../shared/locales.ts";
 import { sendMessage } from "../../shared/messages.ts";
 import { requestPermission } from "../../shared/permissions.ts";
@@ -15,12 +11,16 @@ import {
   getSubscriptionDisplayName,
   isErrorResult,
 } from "../../shared/utilities.ts";
-import { FromNow } from "../from-now.tsx";
+import badgeStyles from "../../styles/badge.module.css";
+import styles from "../../styles/checkbox.module.css";
+import labelStyles from "../../styles/label.module.css";
+import tableStyles from "../../styles/table.module.css";
+import { FromNow } from "../shared/from-now.tsx";
 import { RenameDialog } from "./rename-dialog.tsx";
-import rowStyles from "./row.module.css";
 import { ShowDialog } from "./show-dialog.tsx";
+import localStyles from "./table-row.module.css";
 
-export function ManageSubscription({
+export function SubscriptionTableRow({
   id,
   subscription,
   updating,
@@ -52,7 +52,7 @@ export function ManageSubscription({
           <label className={labelStyles.controlLabel} htmlFor={checkboxId}>
             {getSubscriptionDisplayName(subscription)}
             {subscription.type && subscription.type !== "ruleset" ? (
-              <span className={clsx(badgeStyles.badge, rowStyles.badge)}>
+              <span className={clsx(badgeStyles.badge, localStyles.badge)}>
                 {subscription.type}
               </span>
             ) : null}
