+++ src/scripts/styles/badge.module.css
