+++ src/scripts/styles/label.module.css
