+++ src/scripts/styles/switch.module.css
